requirement.txt
beautifulsoup4==4.11.1
requests==2.27.1
python-dotenv==0.20.0